fix: ESLintの設定を修正
