+++ src/app/components/features/task/CreateTaskModal.tsx
@@ -7,27 +7,27 @@ import {
   DialogHeader,
   DialogTitle,
   DialogDescription,
-} from "@/app/components/Atomic/dialog";
-import { Button } from "@/app/components/Atomic/button";
-import { Input } from "@/app/components/Atomic/input";
-import { Textarea } from "@/app/components/Atomic/textarea";
+} from "@/components/Atomic/dialog";
+import { Button } from "@/components/Atomic/button";
+import { Input } from "@/components/Atomic/input";
+import { Textarea } from "@/components/Atomic/textarea";
 import { useTask } from "@/app/contexts/TaskContext";
-import { Calendar } from "@/app/components/Atomic/calendar";
+import { Calendar } from "@/components/Atomic/calendar";
 import {
   Popover,
   PopoverContent,
   PopoverTrigger,
-} from "@/app/components/Atomic/popover";
+} from "@/components/Atomic/popover";
 import { format } from "date-fns";
 import { CalendarIcon } from "lucide-react";
-import { cn } from "@/app/lib/utils";
+import { cn } from "@/lib/utils";
 import {
   Select,
   SelectContent,
   SelectItem,
   SelectTrigger,
   SelectValue,
-} from "@/app/components/Atomic/select";
+} from "@/components/Atomic/select";
 
 interface CreateTaskModalProps {
   projectId: string;
